completed second exerices, person class now exists and is passed to student and instructor
