+++ person.py
@@ -1,0 +1,6 @@
+class NSSPerson:
+    def __init__(self, firstName, lastName, cohort):
+        self.firstName = firstName
+        self.lastName = lastName
+        self.cohort = cohort
+        self.slackHandle = ""
