+++ student.py
@@ -1,8 +1,7 @@
-class Student:
+from person import NSSPerson
 
-    def __init__(self, firstName, lastName, cohort):
-        self.firstName = firstName
-        self.lastName = lastName
-        self.slackHandle = ""
-        self.cohort = cohort
+class Student(NSSPerson):
+
+    def __init__(self, firstName, lastName, cohort, slackHandle=""):
+        super().__init__(firstName, lastName, cohort)
         self.exercises = set()
